+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v10.2.0 (unreleased)
 ### New features
-- [PR #2661](https://github.com/rqlite/rqlite/pull/2661), [PR #2679](https://github.com/rqlite/rqlite/pull/2679): Support verifying mTLS peer Common Name. Thanks @hifi
+- [PR #2661](https://github.com/rqlite/rqlite/pull/2661), [PR #2679](https://github.com/rqlite/rqlite/pull/2679), [PR #2683](https://github.com/rqlite/rqlite/pull/2683): Support verifying mTLS peer Common Name. Thanks @hifi
 - [PR #2681](https://github.com/rqlite/rqlite/pull/2681): Console supports restore from SQLite data.
   
 ### Implementation changes and bug fixes